+++ src/main/Main.hs
@@ -66,6 +66,7 @@ import           Stack.Fetch
 import           Stack.FileWatch
 import           Stack.Ghci
 import           Stack.Hoogle
+import           Stack.Ls
 import qualified Stack.IDE as IDE
 import qualified Stack.Image as Image
 import           Stack.Init
@@ -299,6 +300,10 @@ commandLineHandler currentDir progName isInterpreter = complicatedOptions
                     "Print out handy path information"
                     pathCmd
                     Stack.Path.pathParser
+        addCommand' "ls"
+                    "List command. (Supports snapshots)"
+                    lsCmd
+                    lsParser
         addCommand' "unpack"
                     "Unpack one or more packages locally"
                     unpackCmd