+++ ChangeLog.md
@@ -10,6 +10,10 @@ Behaviour changes:
 
 Other enhancements:
 
+* A new sub command `ls` has been introduced to stack to view
+  local and remote snapshots present in the system. Use `stack ls
+  snapshots --help` to get more details about it.
+
 Bug fixes:
 
 * For versions of Cabal before 1.24, ensure that the dependencies of